A working AI that will take a king, defend a king, make moves without being prompted, act aggressively, act passively, sacrafice pieces, make trades, and also be stupid at the same time was added with this push.
I apologize for the horrific wrapping that exists inside of the AI class.
I also fixed the formatting errors in the model class, but that's not as bad as the AI class is. I didn't fix the format in the panel class becuase I figured it had been worked on since I last pulled an update.